Update CS scripts

+++ use_case_3/src/R/RPC_models_LASA/RPC_models_cross_sectional_LASA/RPC_models_LASA_CS_overall_model.R
@@ -121,7 +121,7 @@ RPC_models_CS_overall_model <- function(df, config, model = "memory", exclude=c(
       dplyr::left_join(baseline_df[c("id", "date_baseline")], by = "id") %>%
       dplyr::mutate(days_since_baseline = as.numeric(difftime(date, date_baseline, units = "days"))) %>%
       dplyr::filter(days_since_baseline >= 0)
-    df$years_since_baseline <- as.numeric(df$days_since_baseline/365.25, 0)
+    df$years_since_baseline <- as.integer(df$days_since_baseline/365.25, 0)
 
     df <- subset(df, years_since_baseline >= 0)
